refactor(solana): organizes programs's folders better
