+++ packages/solana/programs/adapter/src/contexts/mod.rs
@@ -1,0 +1,3 @@
+pub mod store_hash;
+
+pub use store_hash::*;
