+++ packages/solana/programs/adapter/src/state/mod.rs
@@ -1,0 +1,3 @@
+pub mod hash_account;
+
+pub use hash_account::*;
