+++ packages/solana/programs/hashi/Cargo.toml
@@ -18,3 +18,4 @@ idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
 anchor-lang = "0.30.1"
+adapter = { path = "../adapter", features = ["cpi"] }
