+++ packages/solana/programs/hashi/src/contexts/check_hash_with_threshold.rs
@@ -1,0 +1,4 @@
+use crate::*;
+
+#[derive(Accounts)]
+pub struct CheckHashWithThreshold {}
