+++ packages/solana/programs/hashi/src/contexts/mod.rs
@@ -1,0 +1,3 @@
+pub mod check_hash_with_threshold;
+
+pub use check_hash_with_threshold::*;
